refactor(youtube): create a youtube folder
