+++ lib/pages/section/live/liveCams.dart
@@ -4,8 +4,8 @@ import 'package:get/get.dart';
 import 'package:tv/blocs/live/liveCubit.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tv/controller/textScaleFactorController.dart';
-import 'package:tv/widgets/story/youtubeViewer.dart';
-import 'package:tv/widgets/youtubeLiveViewer.dart';
+import 'package:tv/widgets/youtube/youtubeViewer.dart';
+import 'package:tv/widgets/youtube/youtubeLiveViewer.dart';
 
 class LiveCams extends StatefulWidget {
   @override
