+++ lib/pages/section/show/showStoryPage.dart
@@ -14,7 +14,7 @@ import 'package:tv/models/youtubePlaylistItem.dart';
 import 'package:tv/pages/section/show/showPlaylistTabContent.dart';
 import 'package:tv/services/youtubePlaylistService.dart';
 import 'package:tv/widgets/inlineBannerAdWidget.dart';
-import 'package:tv/widgets/story/youtubePlayer.dart';
+import 'package:tv/widgets/youtube/youtubePlayer.dart';
 
 class ShowStoryPage extends StatefulWidget {
   final String youtubePlayListId;
