+++ lib/widgets/liveWidget.dart
@@ -4,7 +4,7 @@ import 'package:get/get.dart';
 import 'package:tv/blocs/live/liveCubit.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tv/controller/textScaleFactorController.dart';
-import 'package:tv/widgets/youtubeLiveViewer.dart';
+import 'package:tv/widgets/youtube/youtubeLiveViewer.dart';
 
 class LiveWidget extends StatefulWidget {
   final bool needBuildLiveTitle;
